Add support for transaction metrics (#23)
This change introduces a configurable observer listening for transaction
begin and close events. The callbacks will receive additional IO and
allocation statistics collected throughout the execution of the
transaction.

Note: Depending on 'speed' of test, the clock and the clocks accuracy some of
the observer tests might be very flaky, due to duration being 0.
We want to be sure durations taking for the different transaction types.
That's why we're waiting 500ms before critical operations closing a
transaction.