Fix hang queue on open due to panic on atomic op
The atomic transaction counter needs to be aligned to 64bit words, so to
not cause a panic on some architectures (arm or 32bit x86)

Unfortunately the file lock was not released when this panic occured,
making applications hang on startup.

We move the atomic to the top of the file and also ensure the file lock
is correctly released (using defer) if an error or panic occurs on
transaction begin.